feat: 관리자 호출 api는 /admin으로 시작하는 경로로 분리
