+++ src/main/java/roomescape/reservationtime/controller/AdminReservationTimeController.java
@@ -1,0 +1,40 @@
+package roomescape.reservationtime.controller;
+
+import jakarta.validation.Valid;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.reservationtime.entity.ReservationTime;
+import roomescape.reservationtime.payload.ReservationTimeRequest;
+import roomescape.reservationtime.payload.ReservationTimeResponse;
+import roomescape.reservationtime.service.ReservationTimeService;
+
+@RestController
+@RequestMapping("/admin/times")
+public class AdminReservationTimeController {
+
+    private final ReservationTimeService reservationTimeService;
+
+    public AdminReservationTimeController(ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
+    }
+
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public ReservationTimeResponse postTimes(@Valid @RequestBody ReservationTimeRequest request) {
+        ReservationTime reservationTime = reservationTimeService.save(request);
+        return ReservationTimeResponse.from(reservationTime);
+    }
+
+    @DeleteMapping("/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void deleteTimes(@PathVariable Long id) {
+        reservationTimeService.deleteById(id);
+    }
+
+}
