+++ src/main/java/roomescape/theme/controller/AdminThemeController.java
@@ -1,0 +1,39 @@
+package roomescape.theme.controller;
+
+import jakarta.validation.Valid;
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+import roomescape.theme.entity.Theme;
+import roomescape.theme.payload.ThemeRequest;
+import roomescape.theme.payload.ThemeResponse;
+import roomescape.theme.service.ThemeService;
+
+@RestController
+@RequestMapping("/admin/themes")
+public class AdminThemeController {
+
+    private final ThemeService themeService;
+
+    public AdminThemeController(ThemeService themeService) {
+        this.themeService = themeService;
+    }
+
+    @PostMapping
+    @ResponseStatus(HttpStatus.CREATED)
+    public ThemeResponse postTheme(@Valid @RequestBody ThemeRequest request) {
+        Theme theme = themeService.save(request);
+        return ThemeResponse.from(theme);
+    }
+
+    @DeleteMapping("/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void deleteTimes(@PathVariable Long id) {
+        themeService.deleteById(id);
+    }
+}
