+++ src/test/java/roomescape/MissionStep3Test.java
@@ -22,7 +22,7 @@ public class MissionStep3Test {
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(params)
-                .when().post("/times")
+                .when().post("/admin/times")
                 .then().log().all()
                 .statusCode(201);
 
@@ -33,7 +33,7 @@ public class MissionStep3Test {
                 .body("size()", is(1));
 
         RestAssured.given().log().all()
-                .when().delete("/times/1")
+                .when().delete("/admin/times/1")
                 .then().log().all()
                 .statusCode(204);
     }
@@ -46,7 +46,7 @@ public class MissionStep3Test {
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(reservationTime)
-                .when().post("/times")
+                .when().post("/admin/times")
                 .then().log().all()
                 .statusCode(201);
 
@@ -58,7 +58,7 @@ public class MissionStep3Test {
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(theme)
-                .when().post("/themes")
+                .when().post("/admin/themes")
                 .then().log().all()
                 .statusCode(201);
